Paginação Produto e ajuste de Pesquisa de Cidades

+++ tadsGames/src/main/java/br/com/unidospi/Controller/ProdutoController.java
@@ -37,7 +37,7 @@ public class ProdutoController extends HttpServlet {
     public static ArrayList<Produto> listarProdutoPaginado (int offset){
         return ProdutoDAO.listarProdutoPaginado(offset);
     }
-    public static int qtdRegitro(){
+    public static double qtdRegitro(){
         return ProdutoDAO.qtdRegitro();
     }
     public static ArrayList<ProdutoLista> listarProduto (String nomePesquisado,int idEmpresa){
